refactored stockvalueextractor to break into multipleclasses. created utilities folders.
